PHPUnit/Modules/Usage: New - Base schema.
